Kenzie/Sub-MAF-07: Updated comments in classes.
