+++ app/src/main/AndroidManifest.xml
@@ -42,7 +42,7 @@
             android:exported="true" />
 
         <activity
-            android:name="com.terracode.blueharvest.accessibilitySettings.AccessibilitySettingsActivity"
+            android:name="com.terracode.blueharvest.AccessibilitySettingsActivity"
             android:exported="true" />
 
     </application>
